+++ axelor-project/src/main/webapp/project-editor/precache-manifest.126b044f3406a913589f59a8ddbb5144.js
@@ -1,0 +1,30 @@
+self.__precacheManifest = (self.__precacheManifest || []).concat([
+  {
+    "revision": "ebfd85517848746c69e141b0eac20af2",
+    "url": "./index.html"
+  },
+  {
+    "revision": "4c07ffb93de09bbcb42a",
+    "url": "./static/css/2.0b510fe8.chunk.css"
+  },
+  {
+    "revision": "8ca9faad22c970130f38",
+    "url": "./static/css/main.4ed31023.chunk.css"
+  },
+  {
+    "revision": "4c07ffb93de09bbcb42a",
+    "url": "./static/js/2.8c189485.chunk.js"
+  },
+  {
+    "revision": "6dde6af27d388bbd0c95c4e81cf05839",
+    "url": "./static/js/2.8c189485.chunk.js.LICENSE.txt"
+  },
+  {
+    "revision": "8ca9faad22c970130f38",
+    "url": "./static/js/main.752c64eb.chunk.js"
+  },
+  {
+    "revision": "083e6c14fe78135d6547",
+    "url": "./static/js/runtime-main.298acfe2.js"
+  }
+]);